requirements.txt
numpy>=1.22.0
scikit-learn>=1.0.0
sentence-transformers>=2.2.0
flask>=2.0.0
requests>=2.25.0
nltk>=3.6.0
tqdm>=4.60.0
python-dotenv>=0.19.0
langchain
tiktoken